requirements.txt
# TREXIMA v4.0 Production Dependencies
# For SAP BTP Cloud Foundry Deployment

# =============================================================================
# CORE WEB FRAMEWORK
# =============================================================================
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
flask-socketio>=5.3.0

# =============================================================================
# DATABASE
# =============================================================================
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# =============================================================================
# SAP BTP INTEGRATION
# =============================================================================
# XSUAA Authentication
sap-xssec>=4.1.0
python-jose[cryptography]>=3.3.0

# Cloud Foundry Environment
cfenv>=0.5.3

# Object Store (S3-compatible)
boto3>=1.34.0
botocore>=1.34.0

# =============================================================================
# ASYNC & WEBSOCKET
# =============================================================================
gevent>=24.2.0
gevent-websocket>=0.10.1
eventlet>=0.35.0

# =============================================================================
# XML PROCESSING
# =============================================================================
beautifulsoup4>=4.12.0
lxml>=5.0.0

# =============================================================================
# EXCEL PROCESSING
# =============================================================================
openpyxl>=3.1.0

# =============================================================================
# LOCALIZATION
# =============================================================================
babel>=2.14.0

# =============================================================================
# HTTP CLIENT
# =============================================================================
requests>=2.31.0

# =============================================================================
# SAP ODATA CLIENT
# =============================================================================
pyodata>=1.11.0

# =============================================================================
# SECURITY & ENCRYPTION
# =============================================================================
cryptography>=42.0.0

# =============================================================================
# UTILITIES
# =============================================================================
python-dateutil>=2.8.0
uuid>=1.30

# =============================================================================
# PRODUCTION SERVER
# =============================================================================
gunicorn>=21.0.0

# =============================================================================
# OPTIONAL: LOCAL GUI (not needed in BTP)
# =============================================================================
# ttkthemes>=3.2.0
# easygui>=0.98.0
